Renamed system settings to database settings
