+++ Core/Frameworks/BaikalAdmin/Controller/Navigation/Topbar.php
@@ -35,13 +35,13 @@ class Topbar extends \Flake\Core\Controller {
         $oView = new \BaikalAdmin\View\Navigation\Topbar();
 
         $sCurrentRoute = $GLOBALS["ROUTER"]::getCurrentRoute();
-        $sActiveHome = $sActiveUsers = $sActiveSettingsStandard = $sActiveSettingsSystem = "";
+        $sActiveHome = $sActiveUsers = $sActiveSettingsStandard = $sActiveSettingsDatabase = "";
 
         $sControllerForDefaultRoute = $GLOBALS["ROUTER"]::getControllerForRoute("default");
         $sHomeLink = $sControllerForDefaultRoute::link();
         $sUsersLink = \BaikalAdmin\Controller\Users::link();
         $sSettingsStandardLink = \BaikalAdmin\Controller\Settings\Standard::link();
-        $sSettingsSystemLink = \BaikalAdmin\Controller\Settings\System::link();
+        $sSettingsDatabaseLink = \BaikalAdmin\Controller\Settings\Database::link();
         $sLogoutLink = \BaikalAdmin\Controller\Logout::link();
 
         if ($sCurrentRoute === "default") {
@@ -59,18 +59,18 @@ class Topbar extends \Flake\Core\Controller {
             $sActiveSettingsStandard = "active";
         }
 
-        if ($sCurrentRoute === $GLOBALS["ROUTER"]::getRouteForController("\BaikalAdmin\Controller\Settings\System")) {
-            $sActiveSettingsSystem = "active";
+        if ($sCurrentRoute === $GLOBALS["ROUTER"]::getRouteForController("\BaikalAdmin\Controller\Settings\Database")) {
+            $sActiveSettingsDatabase = "active";
         }
 
         $oView->setData("activehome", $sActiveHome);
         $oView->setData("activeusers", $sActiveUsers);
         $oView->setData("activesettingsstandard", $sActiveSettingsStandard);
-        $oView->setData("activesettingssystem", $sActiveSettingsSystem);
+        $oView->setData("activesettingsdatabase", $sActiveSettingsDatabase);
         $oView->setData("homelink", $sHomeLink);
         $oView->setData("userslink", $sUsersLink);
         $oView->setData("settingsstandardlink", $sSettingsStandardLink);
-        $oView->setData("settingssystemlink", $sSettingsSystemLink);
+        $oView->setData("settingsdatabaselink", $sSettingsDatabaseLink);
         $oView->setData("logoutlink", $sLogoutLink);
 
         return $oView->render();
