+++ Core/Frameworks/BaikalAdmin/Resources/Templates/Navigation/Topbar.html
@@ -12,8 +12,8 @@
 				<ul class="nav">
 					<li class="{{ activehome }}"> <a href="{{ homelink }}">Dashboard</a></li>
 					<li class="{{ activeusers }}"> <a href="{{ userslink }}">Users and resources</a></li>
-					<li class="{{ activesettingsstandard }}"> <a href="{{ settingsstandardlink }}">Settings</a></li>
-					<li class="{{ activesettingssystem }}"> <a href="{{ settingssystemlink }}">System settings</a></li>
+					<li class="{{ activesettingsstandard }}"> <a href="{{ settingsstandardlink }}">System Settings</a></li>
+					<li class="{{ activesettingsdatabase }}"> <a href="{{ settingsdatabaselink }}">Database settings</a></li>
 					<li> <a href="{{ logoutlink }}"><i class="icon-eject icon-white"></i> Logout</a></li>
 				</ul>
 			</div>
