+++ Core/Frameworks/BaikalAdmin/Resources/Templates/Settings/Database.html
@@ -1,6 +1,6 @@
 {% autoescape false %}
 <header class="jumbotron subhead" id="overview">
-	<h1><i class="glyph2x-adjust"></i>Baïkal system settings</h1>
+	<h1><i class="glyph2x-adjust"></i>Baïkal database settings</h1>
 </header>
 
 {{ message }}
